move everything into code.ts
